+++ brion/enums.h
@@ -165,11 +165,11 @@ enum TargetType
 /** Classification of neuron substructures. */
 enum SectionType
 {
-    SECTION_SOMA = 1,           //!< neuron cell body
+    SECTION_UNDEFINED = 0,
+    SECTION_SOMA,               //!< neuron cell body
     SECTION_AXON,
     SECTION_DENDRITE,           //!< general or basal dendrite (near to soma)
     SECTION_APICAL_DENDRITE,    //!< apical dendrite (far from soma)
-    SECTION_UNDEFINED
 };
 
 /** The supported attributes of a synapse. */